Removing use of global grunt and adding npm build script
This is a minor modernization of the project. You shouldn't require
users to have global versions of packages installed. Using npm scripts
means the grunt command will be available when the npm script is run.

Note the indentation of package.json has changed to two spaces. This is
because of the use of the automatic command `npm install --save-dev
grunt`. Note any automatic npm command will change the spacing to 2
spaces, so I think it's best to keep it this way.

Additionally this adds missing grunt plugins to the packages, which
presumably you had installed along with your global grunt.