Removing implicit use of THREE and SPE in source
Related to issue #95, the source code of this project relies on the use
of the global SPE variable and the global THREE variable. This makes it
hard to track what files use what dependencies. Plus the javascript
community has moved to commonjs or es6 module syntax for good reason. It
makes it easier to work on projects. Relying on the order of a file
included in an array of strings is not ideal for dependency management.

This also removes the SPE namespace from files which felt redundant, and
also mirrors the elimination of the SPE namespace attachments used
inside those files.

This namespaces all of the objects in a central place src/index.js and
exports that as main from package.json