Fix for `getFromPool()` when pool is empty.
When trying to get an emitter from an empty pool, a new emitter is created when specified, but it is not added to this group like the emitters in `addPool` are. This change makes sure that any emitters created via this method will also automatically add to the group as one would expect.